fix: streak 적용 안되는 버그 수정 - streak 완료여부 계산 Float type으로 타입캐스팅

+++ smudy_backend/backend/user-service/src/main/kotlin/com/ssafy/userservice/service/UserService.kt
@@ -418,7 +418,7 @@ class UserService(
     }
 
     fun isCorrect(total: Int, score: Int): Boolean {
-        return score/total * 100 >= 70
+        return score/total.toFloat() * 100 >= 70
     }
 
     fun isCorrect(average: Int) :  Boolean{
